api/requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
psycopg2==2.7.3.2
pytz==2017.2
django_filter
django_extensions
drf-extensions
factory-boy

python-keystoneclient==3.13.0
python-swiftclient==3.4.0
pandas
flake8
